requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-telegram-bot==21.0.1
